requirements.txt
Brotli==1.0.9
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
dash==1.18.1
dash-auth==1.4.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SeaSurf==0.2.2
future==0.18.2
h11==0.11.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.2.0
phonenumbers==8.12.15
plotly==4.14.1
PyJWT==2.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
retrying==1.3.3
six==1.15.0
twilio==6.50.1
ua-parser==0.10.0
urllib3==1.26.2
Werkzeug==1.0.1
